% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastMultiPCF.R
\name{apply_filter}
\alias{apply_filter}
\title{Applies a given filter over each column of a matrix and
computes the respective sums}
\usage{
apply_filter(x, filter, start, end)
}
\arguments{
\item{x}{The matrix}

\item{filter}{the filter (vector of numeric)}

\item{start}{left-spacing of each row}

\item{end}{right-spacing of each row}
}
\value{
vector of numeric with the filter elements
}
\description{
Applies a given filter over each column of a matrix and
computes the respective sums
}
