% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCNV.R
\name{CCNV}
\alias{CCNV}
\title{Generates the cumulative copy number plot}
\usage{
CCNV(
  mSetsAnno,
  seg_mpcf,
  target_ratios,
  ArrayType,
  colour.amplification,
  colour.loss,
  detail.regions,
  array_type
)
}
\arguments{
\item{mSetsAnno}{A list of the RGSet of the target data, the control data and the annotation data}

\item{seg_mpcf}{a dataframe of the segmentation results of the multi sample segmentation.}

\item{target_ratios}{a dataframe of the intesity of all bins for each sample.}

\item{colour.amplification}{Colour for amplification}

\item{colour.loss}{Colour for loss}

\item{detail.regions}{Either NULL or a vector of gene names.}
}
\value{
returns the plot of the cumulative copy number variation plot
}
\description{
Generates the cumulative copy number plot
}
