% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumul.CNV.R
\name{cumul.CNV}
\alias{cumul.CNV}
\title{compute the cumulative CNV plots}
\usage{
cumul.CNV(
  dataFiles,
  segmentationMode = "all",
  thresh = 0.2,
  gamma = 5,
  colour.amplification = "red3",
  colour.loss = "blue4",
  detail.regions = NULL,
  conumee.version = NULL,
  output = "plot",
  controls = NULL
)
}
\arguments{
\item{dataFiles}{A dataframe with columns ArrayType and Basename}

\item{segmentationMode}{specifying the segmentation mode. Allowed values are single/multi/all.}

\item{thresh}{A positive float (>=0) indicating the threshold for an abberation.}

\item{gamma}{A positive integer (>0) indicating the threshold for the multi-sample segmentation.}

\item{colour.amplification}{Colour for amplification}

\item{colour.loss}{Colour for loss}

\item{detail.regions}{Either NULL or a vector of gene names.
#conumee version new name}

\item{conumee.version}{The version of conumee to use (either 1 or 2). 1 is incompatible with mouse or EPICv2 arrays. NULL will set the version heuristically to 1 for 450k, EPIC and to 2 for Mouse and EPICv2}

\item{output}{determines the type of output. Can be either plot, data or all}

\item{controls}{control samples as a dataframe with columns ArrayType and Basename or NULL. If NULL, publicly available can be used (except for mouse or EPICv2 or mouse data)}
}
\value{
If the output is set to data or all, te segmentation values will be returned, else nothing will be returned and the figures are printed to the default plotting terminal.
}
\description{
compute the cumulative CNV plots
}
