% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastMultiPCF.R
\name{define_cutoff}
\alias{define_cutoff}
\title{Defines a highpass filter with double the specified length}
\usage{
define_cutoff(half_length)
}
\arguments{
\item{half_length}{Parameter for the filter length, which will be 2*half_length}
}
\value{
vector of numeric with the filter elements
}
\description{
Defines a highpass filter with double the specified length
}
