% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatempcf.R
\name{define_sawtooth}
\alias{define_sawtooth}
\title{Defines a sawtooth filter with double the specfied length}
\usage{
define_sawtooth(half_length)
}
\arguments{
\item{half_length}{Parameter for the filter length, which will be 2*half_length}
}
\value{
vector of numeric with the filter elements
}
\description{
Defines a sawtooth filter with double the specfied length
}
