% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumul.CNV.R
\name{get.ArrayType}
\alias{get.ArrayType}
\title{Determines the array type of the user input dataframe}
\usage{
get.ArrayType(dataFiles)
}
\arguments{
\item{dataFiles}{Dataframe with a column labelled "ArrayType" containing the arraytype (450K/EPIC/EPIC2) for each sample.}
}
\value{
A string with either "450K" (only 450K samples), "EPIC" (only EPIC samples), "combined" (both 450K and EPIC samples) and EPIC2 (only EPIC2)
}
\description{
Determines the array type of the user input dataframe
}
