% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.chromAberrations.R
\name{get.chromAberrations}
\alias{get.chromAberrations}
\title{Takes the weighted mean intensity per common segment for each sample or chromosome and determines whether a chromosome
has am amplification (1), a loss (-1) or no change (0). This is loaded into a dataframe and returned. A similarity clustering
is performed using the euclidean distance between the samples to reorder them, based on their similarity. Sample that show similar
segment boundaries (from the combined segmentation) can be processed segment wise (chrom.arm = FALSE) or chromosomal arm wise (chrom.arm = TRUE).
Samples segmented separately will be processed on the choromosomal level.}
\usage{
get.chromAberrations(segmentation_data, thresh = 0.2, chrom.arm = FALSE)
}
\arguments{
\item{segmentation_data}{dataframe that has been returned from the function cumul.CNV}

\item{thresh}{a threshold to determine from when on a segment is considered an aberration}

\item{chrom}{boolean, whether the aberration should be calculated on the segment level (FALSE) or on the chromosomal level (TRUE)}
}
\value{
returns the frequency plot of the samples that were segmented togethher using the multi sample segmentation algorithm
}
\description{
Takes the weighted mean intensity per common segment for each sample or chromosome and determines whether a chromosome
has am amplification (1), a loss (-1) or no change (0). This is loaded into a dataframe and returned. A similarity clustering
is performed using the euclidean distance between the samples to reorder them, based on their similarity. Sample that show similar
segment boundaries (from the combined segmentation) can be processed segment wise (chrom.arm = FALSE) or chromosomal arm wise (chrom.arm = TRUE).
Samples segmented separately will be processed on the choromosomal level.
}
