% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumul.CNV.R
\name{get.ConumeeVersion}
\alias{get.ConumeeVersion}
\title{Determines the conumee version from the array type}
\usage{
get.ConumeeVersion(ArrayType)
}
\arguments{
\item{ArrayType}{The ArrayType as returned by get.ArrayType}
}
\value{
The conumee version (either 1 or 2)
}
\description{
Determines the conumee version from the array type
}
