% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatempcf.R
\name{mark_points}
\alias{mark_points}
\title{Takes a vector of values and marks positions at which the respective elements
are larger than a given quantile}
\usage{
mark_points(mark, sawValue, halflength, frac, start, end)
}
\arguments{
\item{mark}{Vector to write the marks to. Smaller elements are marked with
a zero and larger elements are marked by one}

\item{sawValue}{Vector of numeric. Must have same length as mark}

\item{halflength}{Half the length of the filter used to create sawValue}

\item{frac}{Threshold at 1-frac quantile}

\item{start}{left-spacing of each row}

\item{end}{right-spacing of each row}
}
\value{
the modified mark
}
\description{
Takes a vector of values and marks positions at which the respective elements
are larger than a given quantile
}
