% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSampleSeg.R
\name{multiSampleSeg}
\alias{multiSampleSeg}
\title{Segments the data using the multipcf algorithm and visualizes DNA methylation data and generative cumulative plots using conumee}
\usage{
multiSampleSeg(
  mSetsAnno,
  thresh,
  array_type,
  colour.amplification,
  colour.loss,
  detail.regions,
  showPlot,
  gamma
)
}
\arguments{
\item{mSetsAnno}{A list of the RGSet of the target data, the control data and the annotation data}

\item{thresh}{A positive float (>=0) indicating the threshold for an abberation.}

\item{array_type}{Type of methylation array}

\item{colour.amplification}{Colour for amplification}

\item{colour.loss}{Colour for loss}

\item{detail.regions}{Either NULL or a vector of gene names.}

\item{showPlot}{boolean that determines, if the plots will be displayed}
}
\value{
Nothing. Will print the figures to the default plotting terminal.
}
\description{
Segments the data using the multipcf algorithm and visualizes DNA methylation data and generative cumulative plots using conumee
}
