% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayPlot.R
\name{overlayPlot}
\alias{overlayPlot}
\title{Generates the single sample segmentation aberration frequency plot}
\usage{
overlayPlot(mSetsAnno, segmentation_data, colour.amplification, colour.loss)
}
\arguments{
\item{mSetsAnno}{A list of the RGSet of the target data, the control data and the annotation data}

\item{segmentation_data}{a dataframe of the segmentation results of the single sample segmentation.}

\item{colour.amplification}{Colour for amplification}

\item{colour.loss}{Colour for loss}
}
\value{
returns an overlay plot of all segmentations resulting from the single sample segmentation algorithm
}
\description{
Generates the single sample segmentation aberration frequency plot
}
