% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBinContr.R
\name{sampleBinContr}
\alias{sampleBinContr}
\title{binning target data and publicly available control data and mapping them to
the genome and generating annotation file for bins}
\usage{
sampleBinContr(target_rgset, ArrayType, controls)
}
\arguments{
\item{target_rgset}{RGchannelSet of data to be binned}

\item{ArrayType}{character representing a that Array Type used for acquiring
data}
}
\value{
A list containing three objects, the mapped MethylSet of the target
data, the mapped MethylSet of the control Data and the annotation file of
the bins.
}
\description{
binning target data and publicly available control data and mapping them to
the genome and generating annotation file for bins
}
