% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum.CNV.R
\name{segment.Plot}
\alias{segment.Plot}
\title{Segments and visualizes DNA methylation data and generative cumulative plots}
\usage{
segment.Plot(
  target_rgset,
  array_type,
  segmentationMode,
  gamma,
  thresh,
  colour.amplification,
  colour.loss,
  detail.regions,
  conumee.version
)
}
\arguments{
\item{segmentationMode}{specifying the segmentation mode. Allowed values are single/multi/all.}

\item{gamma}{A positive integer (>0) indicating the threshold for the multi-sample segmentation.}

\item{thresh}{A positive float (>=0) indicating the threshold for an abberation.}

\item{colour.amplification}{Colour for amplification}

\item{colour.loss}{Colour for loss}

\item{detail.regions}{Either NULL or a vector of gene names.}

\item{conumee.version}{The conumee version to use.}

\item{dataFiles}{A dataframe with columns ArrayType and Basename}
}
\value{
Returns the segmentation values either as a dataframe or a list of two dataframes.
}
\description{
Segments and visualizes DNA methylation data and generative cumulative plots
}
