% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleFrequencyPlot.R
\name{singleFrequencyPlot}
\alias{singleFrequencyPlot}
\title{Generates the single sample segmentation aberration frequency plot}
\usage{
singleFrequencyPlot(
  mSetsAnno,
  segmentation_data,
  colour.amplification,
  colour.loss,
  thresh
)
}
\arguments{
\item{mSetsAnno}{A list of the RGSet of the target data, the control data and the annotation data}

\item{segmentation_data}{a dataframe of the segmentation results of the single sample segmentation.}

\item{colour.amplification}{Colour for amplification}

\item{colour.loss}{Colour for loss}

\item{thresh}{a float specifying the threshold when a segment is called as an aberration}

\item{target_ratios}{a dataframe of the intesity of all bins for each sample.}
}
\value{
returns the frequency plot of the samples that were segmented together using the single sample segmentation algorithm
}
\description{
Generates the single sample segmentation aberration frequency plot
}
