% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSampleSeg.R
\name{singleSampleSeg}
\alias{singleSampleSeg}
\title{Segments the data using the conumee package and visualizes DNA methylation data and generative cumulative plots}
\usage{
singleSampleSeg(mSetsAnno, thresh, colour.amplification, colour.loss)
}
\arguments{
\item{mSetsAnno}{A list of the RGSet of the target data, the control data and the annotation data}

\item{thresh}{A positive float (>=0) indicating the threshold for an abberation.}

\item{colour.amplification}{Colour for amplification}

\item{colour.loss}{Colour for loss}

\item{detail.regions}{Either NULL or a vector of gene names.}
}
\value{
Nothing. Will print the figures to the default plotting terminal.
}
\description{
Segments the data using the conumee package and visualizes DNA methylation data and generative cumulative plots
}
