% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatempcf.R
\name{updatempcf}
\alias{updatempcf}
\title{Segment multiple data together using the partial least squares regression
as presented in XXX}
\usage{
updatempcf(data, gamma = 5, frac1 = 0.15, frac2 = 0.15)
}
\arguments{
\item{data}{dataframe of binned copy number data in the format (samples,
values) with two additional columns. The first representing the chromosome,
the second the position on the chromosome}

\item{gamma}{value representing a threshhold for the segmentation}
}
\value{
A dataframe representing segmentation data, containing columns for
the chromosome ,segmentation start ,segmentation end ,number of bins in that
segment, mean value for that segment, each segmentation value for each sample
}
\description{
Segment multiple data together using the partial least squares regression
as presented in XXX
}
