% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastMultiPCF.R
\name{mark_points}
\alias{mark_points}
\title{Takes a vector of values and marks positions at which the respective elements
are larger than a given quantile}
\usage{
mark_points(bpoints, hpfValue, halflength, frac, start, end)
}
\arguments{
\item{bpoints}{Vector to write the marks to. Smaller elements are marked with
a zero and larger elements are marked by one}

\item{hpfValue}{Vector of numeric. Must have same length as bpoints}

\item{halflength}{Half the length of the filter used to create hpfValue}

\item{frac}{Threshold at 1-frac quantile}

\item{start}{left-spacing of each row}

\item{end}{right-spacing of each row}
}
\value{
the modified vector of potential breakpoints
}
\description{
Takes a vector of values and marks positions at which the respective elements
are larger than a given quantile
}
